+++ src/models/tags.model.js
@@ -2,7 +2,7 @@ const db = require('../lib/db.lib')
 const { isExist, isStringExist, updateColumn } = require('../moduls/handling')
 
 
-exports.findAll = async () => {
+exports.findAll = async (searchKey='', sortBy="id", order="ASC", page=1) => {
     const orderType = ["ASC", "DESC"]
     order = orderType.includes(order)? order : "ASC"
     
@@ -20,19 +20,22 @@ exports.findAll = async () => {
         })
     
         const sql = `
-        SELECT "id", "name", "createdAt"
+        SELECT *
         FROM "tags" WHERE "name" ILIKE $1
         ORDER BY ${columnSort.join(', ')}
         LIMIT ${limit} OFFSET ${offset}
         `
         const values = [`%${searchKey}%`]
         const {rows} = await db.query(sql, values)
+        if(!rows.length){
+            throw new Error(`no data found`)
+        }
         return rows
     }
 
     const sql = `
-    SELECT "id", "name","createdAt"
-    FROM "tags" WHERE "name" ilike $1
+    SELECT *
+    FROM "tags" WHERE "name" ILIKE $1
     ORDER BY "${sortBy}" ${order}
     LIMIT ${limit} OFFSET ${offset}
     `
@@ -40,13 +43,16 @@ exports.findAll = async () => {
     console.log(sql)
     const values = [`%${searchKey}%`]
     const {rows} = await db.query(sql, values)
+    if(!rows.length){
+        throw new Error(`no data found`)
+    }
     return rows
 }
 
 
 exports.findOne = async (id) => {                                                                             // mencari data berdasarkan column dengan constraint unique
     const sql = `
-    SELECT "id", "name", "createdAt"
+    SELECT *
     FROM "tags" WHERE "id" = $1
     `
     const  values = [id]
@@ -59,7 +65,7 @@ exports.findOne = async (id) => {                                               
 
 
 exports.insert = async (body) => {
-    const queryString = await isStringExist("tags", "name", data.name)
+    const queryString = await isStringExist("tags", "name", body.name)
     if(queryString){
         throw new Error(queryString)
     }
@@ -73,6 +79,10 @@ exports.insert = async (body) => {
 
 
 exports.update = async (id, body) => {
+    if(isNaN(id)){
+        throw new Error(`invalid input`)
+    }
+
     const queryId = await isExist("tags", id)
     if(queryId){
         throw new Error(queryId)
@@ -84,15 +94,20 @@ exports.update = async (id, body) => {
     }
         
 
-    return await updateColumn(id, body, "users")
+    return await updateColumn(id, body, "tags")
 }
 
 
 exports.delete = async (id) => {
+    if(isNaN(id)){
+        throw new Error(`invalid input`)
+    }
+
     const queryId = await isExist("tags", id)
     if(queryId){
         throw new Error(queryId)
     }
+    
     const sql = `DELETE FROM "tags" WHERE "id" = $1 RETURNING *`
     const values = [id]
     const {rows} = await db.query(sql, values)
